Move type definitions to separate module
